requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Conservative Django
django==1.11.10 # pyup: <2.0

# Configuration
django-environ==0.4.4



# Forms
django-crispy-forms==1.7.0

# Models
django-model-utils==3.1.1

# Images
Pillow==5.0.0

# Password storage
argon2-cffi==18.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.35.0

# On Windows, you must download/install psycopg2 manually
# from http://www.lfd.uci.edu/~gohlke/pythonlibs/#psycopg


# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2018.3

# Redis support
django-redis==4.8.0
redis>=2.10.5


celery==3.1.25 # pyup: <4.0


# Python-PostgreSQL Database Adapter
# If using Win for dev, this assumes Unix in test/prod
psycopg2==2.7.4

# Your custom requirements go here
py-solc==2.1.0
